Moving JavaScript code out of main module
