+++ TrainingExtensions/torch/src/python/aimet_torch/v2/visualization_tools/quant_stats_visualization.py
@@ -39,6 +39,7 @@
 
 
 import os
+from pathlib import Path
 import torch
 from bokeh.events import DocumentReady, Reset
 from bokeh.layouts import row, column
@@ -126,7 +127,7 @@ def visualize_stats(sim: QuantizationSimModel, dummy_input, save_path: str = Non
     visualizer = QuantStatsVisualizer(stats_dict)
 
     # Save an interactive bokeh plot as a standalone html
-    visualizer.export_plot_as_html(save_path, mode="basic_stats")
+    visualizer.export_plot_as_html(save_path)
 
 
 def visualize_advanced_stats(sim: QuantizationSimModel, dummy_input, save_path: str = "./quant_stats_visualization.html") -> None:
@@ -177,7 +178,7 @@ def visualize_advanced_stats(sim: QuantizationSimModel, dummy_input, save_path: 
     visualizer = QuantStatsVisualizer(stats_dict)
 
     # Save an interactive bokeh plot as a standalone html
-    visualizer.export_plot_as_html(save_path, mode="advanced_stats")
+    visualizer.export_plot_as_html(save_path)
 
 
 def _check_path(path: str):
@@ -570,103 +571,7 @@ class QuantStatsVisualizer:
             max_thresh_filter=tableobjects.filters.max_thresh_filter,
             name_filter=tableobjects.filters.name_filter,
             select=inputwidgets.table_view_select,
-        ), code="""
-                // Function to adaptively format numerical values in scientific notation
-                // if they are large in magnitude
-                function formatValue(value) {
-                    if (Math.abs(value) < 1e-3 || Math.abs(value) > 1e3) {
-                        return value.toExponential(2);
-                    } else {
-                        return value.toFixed(2);
-                    }
-                }
-
-                function booleanAnd(arr1, arr2) {
-                    return arr1.map((value, index) => value && arr2[index]);
-                }
-
-                function booleanOr(arr1, arr2) {
-                    return arr1.map((value, index) => value || arr2[index]);
-                }
-
-                function findMin(a, b) {
-                    if (a<=b) {
-                        return a;
-                    }
-                    return b;
-                }
-
-                function findMax(a, b) {
-                    if (a>=b) {
-                        return a;
-                    }
-                    return b;
-                }
-
-
-                // Reading values from input widgets and setting plot y axis range  
-                const limits_data = limits_source.data;
-                limits_data['ymax'] = [parseFloat(ymax_input.value)];
-                limits_data['ymin'] = [parseFloat(ymin_input.value)];
-                plot.y_range.start = limits_data['ymin'][0]*1.05;
-                plot.y_range.end = limits_data['ymax'][0]*1.05;
-                limits_data['maxclip'] = [parseFloat(maxclip_input.value)];
-                limits_data['minclip'] = [parseFloat(minclip_input.value)];
-
-                const source_data = data_source.data;
-                const idx = source_data['idx'];
-                const minlist = source_data['minlist'];
-                const maxlist = source_data['maxlist'];
-                const namelist = source_data['namelist'];
-                source_data['marker_yminlist'] = source_data['minlist'].map(t => findMax(t, limits_data['ymin'][0]));
-                source_data['marker_ymaxlist'] = source_data['maxlist'].map(t => findMin(t, limits_data['ymax'][0]));
-
-                // Updating the filters for finding layers that cross the min or max thresholds
-                min_thresh_filter.booleans = minlist.map(t => t <= limits_data['minclip'][0]);
-                max_thresh_filter.booleans = maxlist.map(t => t >= limits_data['maxclip'][0]);
-
-                let table_booleans;
-                const table_idx = [];
-                const table_namelist = [];
-                const table_minlist = [];
-                const table_maxlist = [];
-
-                var view = select.value;
-                if (view == "All") {
-                    table_booleans = name_filter.booleans;
-                } else if (view == "Min") {
-                    table_booleans = booleanAnd(name_filter.booleans, min_thresh_filter.booleans);
-                } else if (view == "Max") {
-                    table_booleans = booleanAnd(name_filter.booleans, max_thresh_filter.booleans);
-                } else if (view == "Min | Max") {
-                    table_booleans = booleanAnd(name_filter.booleans, booleanOr(min_thresh_filter.booleans, max_thresh_filter.booleans));
-                } else if (view == "Min & Max") {
-                    table_booleans = booleanAnd(name_filter.booleans, booleanAnd(min_thresh_filter.booleans, max_thresh_filter.booleans));
-                }
-
-                for (let i = 0; i < idx.length; i++) {
-                    if (table_booleans[i] == true) {
-                        table_idx.push(idx[i]);
-                        table_namelist.push(namelist[i]);
-                        table_minlist.push(minlist[i]);
-                        table_maxlist.push(maxlist[i]);
-                    }
-                }
-
-                table_data_source.data["idx"] = table_idx;
-                table_data_source.data["namelist"] = table_namelist;
-                table_data_source.data["minlist"] = table_minlist;
-                table_data_source.data["maxlist"] = table_maxlist;
-
-                selected_data_source.data["floor"].push(limits_source.data['ymin'][0]*1.05);
-                selected_data_source.data["ceil"].push(limits_source.data['ymax'][0]*1.05);  
-
-                // Emitting the changes made to ColumnDataSources
-                limits_source.change.emit();
-                data_source.change.emit();
-                table_data_source.change.emit();
-                selected_data_source.change.emit();
-            """)
+        ), code=(Path(__file__).parent / "quant_stats_visualization_JS_code/limit_change_callback.js").read_text("utf8"))
 
         customcallbacks.reset_callback = CustomJS(args=dict(
             limits_source=datasources.limits_source,
@@ -686,125 +591,7 @@ class QuantStatsVisualizer:
             min_thresh_filter=tableobjects.filters.min_thresh_filter,
             max_thresh_filter=tableobjects.filters.max_thresh_filter,
             name_filter=tableobjects.filters.name_filter,
-        ), code="""
-                // Function to adaptively format numerical values in scientific notation
-                // if they are large in magnitude
-                function formatValue(value) {
-                    if (Math.abs(value) < 1e-3 || Math.abs(value) > 1e3) {
-                        return value.toExponential(2);
-                    } else {
-                        return value.toFixed(2);
-                    }
-                }
-
-                function booleanAnd(arr1, arr2) {
-                    return arr1.map((value, index) => value && arr2[index]);
-                }
-
-                function booleanOr(arr1, arr2) {
-                    return arr1.map((value, index) => value || arr2[index]);
-                }
-
-                function findMin(a, b) {
-                    if (a<=b) {
-                        return a;
-                    }
-                    return b;
-                }
-
-                function findMax(a, b) {
-                    if (a>=b) {
-                        return a;
-                    }
-                    return b;
-                }
-
-                // Resetting the limits source with default values
-                limits_source.data['ymax'] = default_values_source.data['default_ymax'];
-                limits_source.data['ymin'] = default_values_source.data['default_ymin'];
-                limits_source.data['xmax'] = default_values_source.data['default_xmax'];
-                limits_source.data['xmin'] = default_values_source.data['default_xmin'];
-                limits_source.data['maxclip'] = default_values_source.data['default_maxclip'];
-                limits_source.data['minclip'] = default_values_source.data['default_minclip'];
-                const limits_data = limits_source.data;
-
-                // Resetting the plot ranges
-                plot.y_range.start = limits_data['ymin'][0]*1.05;
-                plot.y_range.end = limits_data['ymax'][0]*1.05;
-                plot.x_range.start = limits_data['xmin'][0];
-                plot.x_range.end = limits_data['xmax'][0];
-
-                // Resetting the input widget values
-                ymax_input.value = limits_data['ymax'][0].toString();
-                ymin_input.value = limits_data['ymin'][0].toString();
-                maxclip_input.value = limits_data['maxclip'][0].toString();
-                minclip_input.value = limits_data['minclip'][0].toString();
-
-                const source_data = data_source.data;
-                const idx = source_data['idx'];
-                const minlist = source_data['minlist'];
-                const maxlist = source_data['maxlist'];
-                const namelist = source_data['namelist'];
-
-                source_data['marker_yminlist'] = source_data['minlist'].map(t => findMax(t, limits_data['ymin'][0]));
-                source_data['marker_ymaxlist'] = source_data['maxlist'].map(t => findMin(t, limits_data['ymax'][0]));
-
-                min_thresh_filter.booleans = minlist.map(t => t <= limits_data['minclip'][0]);
-                max_thresh_filter.booleans = maxlist.map(t => t >= limits_data['maxclip'][0]);
-
-                name_filter.booleans = Array(idx.length).fill(true);
-                name_input.value = "";
-
-                let table_booleans;
-                const table_idx = [];
-                const table_namelist = [];
-                const table_minlist = [];
-                const table_maxlist = [];
-
-                select.value = "Min | Max";
-                var view = select.value;
-                if (view == "All") {
-                    table_booleans = name_filter.booleans;
-                } else if (view == "Min") {
-                    table_booleans = booleanAnd(name_filter.booleans, min_thresh_filter.booleans);
-                } else if (view == "Max") {
-                    table_booleans = booleanAnd(name_filter.booleans, max_thresh_filter.booleans);
-                } else if (view == "Min | Max") {
-                    table_booleans = booleanAnd(name_filter.booleans, booleanOr(min_thresh_filter.booleans, max_thresh_filter.booleans));
-                } else if (view == "Min & Max") {
-                    table_booleans = booleanAnd(name_filter.booleans, booleanAnd(min_thresh_filter.booleans, max_thresh_filter.booleans));
-                }
-
-                for (let i = 0; i < idx.length; i++) {
-                    if (table_booleans[i] == true) {
-                        table_idx.push(idx[i]);
-                        table_namelist.push(namelist[i]);
-                        table_minlist.push(minlist[i]);
-                        table_maxlist.push(maxlist[i]);
-                    }
-                }
-
-                table_data_source.data["idx"] = table_idx;
-                table_data_source.data["namelist"] = table_namelist;
-                table_data_source.data["minlist"] = table_minlist;
-                table_data_source.data["maxlist"] = table_maxlist;
-
-                table_data_source.selected.indices = [];
-                data_source.data["selected"] = Array(data_source.data["idx"].length).fill(false);
-
-                selected_data_source.data["idx"] = [];
-                selected_data_source.data["namelist"] = [];
-                selected_data_source.data["floor"] = [];
-                selected_data_source.data["ceil"] = [];
-                selected_data_source.data["minlist"] = [];
-                selected_data_source.data["maxlist"] = [];
-
-                // Emitting the changes made to ColumnDataSources
-                limits_source.change.emit();
-                data_source.change.emit();
-                table_data_source.change.emit();
-                selected_data_source.change.emit();
-            """)
+        ), code=(Path(__file__).parent / "quant_stats_visualization_JS_code/reset_callback.js").read_text("utf8"))
 
         customcallbacks.name_filter_callback = CustomJS(args=dict(
             data_source=datasources.data_source,
@@ -814,74 +601,7 @@ class QuantStatsVisualizer:
             max_thresh_filter=tableobjects.filters.max_thresh_filter,
             name_filter=tableobjects.filters.name_filter,
             select=inputwidgets.table_view_select,
-        ), code="""
-                function booleanAnd(arr1, arr2) {
-                    return arr1.map((value, index) => value && arr2[index]);
-                }
-
-                function booleanOr(arr1, arr2) {
-                    return arr1.map((value, index) => value || arr2[index]);
-                }
-                // Filter all names having entered pattern as a substring
-                name_filter.booleans = Array.from(data_source.data['namelist']).map(t => t.includes(cb_obj.value));
-
-                const limits_data = limits_source.data;
-                const source_data = data_source.data;
-                const idx = source_data['idx'];
-                const minlist = source_data['minlist'];
-                const maxlist = source_data['maxlist'];
-                const namelist = source_data['namelist'];
-
-                let table_booleans;
-                const table_idx = [];
-                const table_namelist = [];
-                const table_minlist = [];
-                const table_maxlist = [];
-
-                var view = select.value;
-                if (view == "All") {
-                    table_booleans = name_filter.booleans;
-                } else if (view == "Min") {
-                    table_booleans = booleanAnd(name_filter.booleans, min_thresh_filter.booleans);
-                } else if (view == "Max") {
-                    table_booleans = booleanAnd(name_filter.booleans, max_thresh_filter.booleans);
-                } else if (view == "Min | Max") {
-                    table_booleans = booleanAnd(name_filter.booleans, booleanOr(min_thresh_filter.booleans, max_thresh_filter.booleans));
-                } else if (view == "Min & Max") {
-                    table_booleans = booleanAnd(name_filter.booleans, booleanAnd(min_thresh_filter.booleans, max_thresh_filter.booleans));
-                }
-
-                for (let i = 0; i < idx.length; i++) {
-                    if (table_booleans[i] == true) {
-                        table_idx.push(idx[i]);
-                        table_namelist.push(namelist[i]);
-                        table_minlist.push(minlist[i]);
-                        table_maxlist.push(maxlist[i]);
-                    }
-                }
-
-                table_data_source.data["idx"] = table_idx;
-                table_data_source.data["namelist"] = table_namelist;
-                table_data_source.data["minlist"] = table_minlist;
-                table_data_source.data["maxlist"] = table_maxlist; 
-
-                table_data_source.change.emit();
-
-                const selected_indices = [];
-                var layer_idx;
-                for (let i = 0; i < table_idx.length; i++) {
-                    layer_idx = table_idx[i];
-                    if (data_source.data["selected"][layer_idx] == true) {
-                        selected_indices.push(i);
-                    }
-                }
-
-                table_data_source.selected.indices = selected_indices;
-
-                table_data_source.change.emit();
-                table.name = "placeholder_1";
-                table.name = "placeholder_0";
-            """)
+        ), code=(Path(__file__).parent / "quant_stats_visualization_JS_code/name_filter_callback.js").read_text("utf8"))
 
         customcallbacks.select_table_view_callback = CustomJS(args=dict(
             data_source=datasources.data_source,
@@ -891,109 +611,14 @@ class QuantStatsVisualizer:
             max_thresh_filter=tableobjects.filters.max_thresh_filter,
             name_filter=tableobjects.filters.name_filter,
             table=tableobjects.data_table
-        ), code="""
-                function booleanAnd(arr1, arr2) {
-                    return arr1.map((value, index) => value && arr2[index]);
-                }
-
-                function booleanOr(arr1, arr2) {
-                    return arr1.map((value, index) => value || arr2[index]);
-                }
-
-                const source_data = data_source.data;
-                const idx = source_data['idx'];
-                const minlist = source_data['minlist'];
-                const maxlist = source_data['maxlist'];
-                const namelist = source_data['namelist'];
-
-                let table_booleans;
-                const table_idx = [];
-                const table_namelist = [];
-                const table_minlist = [];
-                const table_maxlist = [];
-
-                var view = select.value;
-                if (view == "All") {
-                    table_booleans = name_filter.booleans;
-                } else if (view == "Min") {
-                    table_booleans = booleanAnd(name_filter.booleans, min_thresh_filter.booleans);
-                } else if (view == "Max") {
-                    table_booleans = booleanAnd(name_filter.booleans, max_thresh_filter.booleans);
-                } else if (view == "Min | Max") {
-                    table_booleans = booleanAnd(name_filter.booleans, booleanOr(min_thresh_filter.booleans, max_thresh_filter.booleans));
-                } else if (view == "Min & Max") {
-                    table_booleans = booleanAnd(name_filter.booleans, booleanAnd(min_thresh_filter.booleans, max_thresh_filter.booleans));
-                }
-
-                for (let i = 0; i < idx.length; i++) {
-                    if (table_booleans[i] == true) {
-                        table_idx.push(idx[i]);
-                        table_namelist.push(namelist[i]);
-                        table_minlist.push(minlist[i]);
-                        table_maxlist.push(maxlist[i]);
-                    }
-                }
-
-                table_data_source.data["idx"] = table_idx;
-                table_data_source.data["namelist"] = table_namelist;
-                table_data_source.data["minlist"] = table_minlist;
-                table_data_source.data["maxlist"] = table_maxlist;
-
-                table_data_source.change.emit();
-
-                const selected_indices = [];
-                var layer_idx; 
-                for (let i = 0; i < table_idx.length; i++) {
-                    layer_idx = table_idx[i];
-                    if (data_source.data["selected"][layer_idx] == true) {
-                        selected_indices.push(i);
-                    }
-                }
-
-                table_data_source.selected.indices = selected_indices;
-
-                table_data_source.change.emit();
-
-                // Force redraw of table by making an inert change to table properties
-                table.name = "placeholder_1";
-                table.name = "placeholder_0";
-        """)
+        ), code=(Path(__file__).parent / "quant_stats_visualization_JS_code/select_table_view_callback.js").read_text("utf8"))
 
         customcallbacks.table_selection_callback = CustomJS(args=dict(
             data_source=datasources.data_source,
             table_data_source=datasources.table_data_source,
             selected_data_source=datasources.selected_data_source,
             limits_source=datasources.limits_source,
-        ), code="""
-                table_data_source.data["idx"].forEach(i => {
-                    data_source.data["selected"][i] = false;
-                });
-                table_data_source.selected.indices.forEach(i => {
-                    let layer_idx = table_data_source.data["idx"][i];
-                    data_source.data["selected"][layer_idx] = true;
-                });
-                data_source.change.emit();
-
-                selected_data_source.data["namelist"] = [];
-                selected_data_source.data["idx"] = [];
-                selected_data_source.data["floor"] = [];
-                selected_data_source.data["ceil"] = [];
-                selected_data_source.data["minlist"] = [];
-                selected_data_source.data["maxlist"] = [];
-                selected_data_source.change.emit();
-
-                data_source.data["selected"].forEach((bool,index) => {
-                    if (bool==true) {
-                        selected_data_source.data["namelist"].push(data_source.data["namelist"][index]);
-                        selected_data_source.data["idx"].push(index);
-                        selected_data_source.data["floor"].push(limits_source.data['ymin'][0]*1.05);
-                        selected_data_source.data["ceil"].push(limits_source.data['ymax'][0]*1.05);
-                        selected_data_source.data["minlist"].push(data_source.data["minlist"][index]);
-                        selected_data_source.data["maxlist"].push(data_source.data["maxlist"][index]);
-                    }
-                })
-                selected_data_source.change.emit();
-        """)
+        ), code=(Path(__file__).parent / "quant_stats_visualization_JS_code/table_selection_callback.js").read_text("utf8"))
 
         return customcallbacks
 
@@ -1022,12 +647,11 @@ class QuantStatsVisualizer:
 
         return layout
 
-    def export_plot_as_html(self, save_path: str, mode: str) -> None:
+    def export_plot_as_html(self, save_path: str) -> None:
         """
         Method for constructing the visualization and saving it to the given path.
 
         :param save_path: Path for saving the visualization.
-        :param mode: Whether to plot basic stats or advanced stats.
         """
 
         curdoc().theme = 'light_minimal'
